+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Add support for checking that the value is included_in
+
 ## [0.3.0] - 2016-11-09
 ### Added
 - Add support for float, decimal, bool, date, time, date_time, array, hash